refactor and make body_view
